update check before the sift link

+++ ppml/trusted-big-data-ml/python/docker-graphene/init.sh
@@ -7,9 +7,16 @@ sgx_mem_size=$SGX_MEM_SIZE
 
 make SGX=1 GRAPHENEDIR=/graphene THIS_DIR=/ppml/trusted-big-data-ml  SPARK_LOCAL_IP=$local_ip SPARK_USER=root G_SGX_SIZE=$sgx_mem_size
 
-cd /dev
-mkdir sgx
-ln -s /dev/sgx_enclave /dev/sgx/enclave
-ln -s /dev/sgx_provision /dev/sgx/provision
+if [ ! -f "/dev/sgx/enclave" ];then
+ cd /dev
+ mkdir sgx
+ ln -s /dev/sgx_enclave /dev/sgx/enclave
+ cd /ppml/trusted-big-data-ml
+fi
 
-cd /ppml/trusted-big-data-ml
+if [ ! -f "/dev/sgx/provision" ];then
+ cd /dev
+ mkdir sgx
+ ln -s /dev/sgx_provision /dev/sgx/provision
+ cd /ppml/trusted-big-data-ml
+fi
